@@ -18,7 +18,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -26,18 +26,20 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.4
 
-  google_fonts: ^2.0.0
-  flutter_svg: ^0.22.0
-  fl_chart: ^0.35.0
-  provider: ^5.0.0
-  image_picker: ^0.8.4+11
+  google_fonts: ^2.3.2
+  flutter_svg: ^1.0.3
+  fl_chart: ^0.50.6
+  provider: ^6.0.2
+  image_picker: ^0.8.4
   flutter_animated_dialog: ^2.0.1
-  simple_animations: ^1.3.3
-  dio: ^4.0.4
-  shared_preferences: ^0.5.6+1
-  searchfield: ^0.5.6
+  simple_animations: ^4.0.2
+  dio: ^4.0.6
+  shared_preferences: ^2.0.15
+  searchfield: ^0.6.5
+  universal_io: ^2.0.4
+  editable: ^2.0.0
 
 dev_dependencies:
   flutter_test:
