@@ -4,7 +4,7 @@ import 'package:admin/screens/main/main_screen.dart';
 import 'package:flutter/material.dart';
 
 class Raccourcis {
-  final String imgSrc, title;
+  final String? imgSrc, title;
   // final Function() onTapAction; // replace Function() with VoidCallback?
 
   Raccourcis({this.imgSrc, this.title, info});
